+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
+## [29.0.0] - 2026-02-08
 ### Added
 - Add pending and skipped via exceptions ([#238](https://github.com/cucumber/compatibility-kit/pull/238))
 - Samples for status escalations ([#239](https://github.com/cucumber/compatibility-kit/pull/239))
@@ -345,7 +347,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - First release
 
-[Unreleased]: https://github.com/cucumber/compatibility-kit/compare/v28.0.1...HEAD
+[Unreleased]: https://github.com/cucumber/compatibility-kit/compare/v29.0.0...HEAD
+[29.0.0]: https://github.com/cucumber/compatibility-kit/compare/v28.0.1...v29.0.0
 [28.0.1]: https://github.com/cucumber/compatibility-kit/compare/v28.0.0...v28.0.1
 [28.0.0]: https://github.com/cucumber/compatibility-kit/compare/v27.0.0...v28.0.0
 [27.0.0]: https://github.com/cucumber/compatibility-kit/compare/v26.0.0...v27.0.0